SOLVED ERROR HANDLE IN GULP PIPE
the problem was that
errors thrown from somewhere in gulp pipe, it stops all the remain
process, and gulp exits with error message.

but there was another problem event if i catch this error using plumber
or custum error block emitting 'end'. it stops all pipeline including
watch.

in short i had to catch exceptions, and make flow of tasks in gulp
pipeline.

solution was so easy.
sepating browserify and gulp was that.
when error thrown, browserify returns previously build outputs and if
block inject it into pipe.

thats it.

but i spent almost 5 hours to solve this...